+++ test/id-api.test.ts
@@ -1,8 +1,9 @@
 import keypair from 'keypair';
 import nock from 'nock';
 import * as packageJson from '../package.json';
+import businessVerificationResp from './fixtures/business_verification_response.json';
 
-import { IDApi, Signature } from '..';
+import { IDApi, Signature, JOB_TYPE } from '..';
 
 const pair = keypair();
 
@@ -86,7 +87,7 @@ describe('IDapi', () => {
       const instance = new IDApi('001', Buffer.from(pair.public).toString('base64'), 0);
       const partner_params = { user_id: '1', job_id: '1', job_type: 4 };
       // @ts-ignore
-      await expect(instance.submit_job(partner_params, null)).rejects.toThrow(new Error('Please ensure that you are setting your job_type to 5 to query ID Api'));
+      await expect(instance.submit_job(partner_params, null)).rejects.toThrow(new Error('Please ensure that you are setting your job_type to 5 or 7 to query ID Api'));
     });
 
     it('should be able to send a job', async () => {
@@ -191,6 +192,68 @@ describe('IDapi', () => {
       // err.message in this case should be '2204:unauthorized'
       expect(error.message).toBeUndefined();
       expect(response).toEqual(undefined);
+      expect(scope.isDone()).toBe(true);
+    });
+  });
+
+  describe('business_verification', () => {
+    it('successfully sends a business verification job', async () => {
+      expect.assertions(3);
+      const instance = new IDApi('001', 'api_key', 0);
+      const partner_params = {
+        user_id: '1',
+        job_id: '1',
+        job_type: JOB_TYPE.BUSINESS_VERIFICATION,
+      };
+      const id_info = {
+        country: 'NG',
+        id_type: 'BUSINESS_REGISTRATION',
+        id_number: 'A000000',
+        business_type: 'co',
+      };
+
+      const postMock = jest.fn(() => true);
+      const scope = nock('https://testapi.smileidentity.com')
+        .post('/v1/business_verification', postMock)
+        .reply(200, businessVerificationResp.success);
+
+      const resp = await instance.submit_job(partner_params, id_info);
+      expect(resp.data).toEqual(businessVerificationResp.success);
+      expect(scope.isDone()).toBe(true);
+      expect(postMock).toHaveBeenCalledWith({
+        api_key: 'api_key',
+        business_type: id_info.business_type,
+        country: id_info.country,
+        id_number: id_info.id_number,
+        id_type: id_info.id_type,
+        partner_id: '001',
+        timestamp: expect.any(String),
+        signature: expect.any(String),
+        partner_params,
+      });
+    });
+
+    it('report an error on unsuccessfull business verification', async () => {
+      expect.assertions(2);
+      const instance = new IDApi('001', 'api_key', 0);
+      const partner_params = {
+        user_id: '1',
+        job_id: '1',
+        job_type: JOB_TYPE.BUSINESS_VERIFICATION,
+      };
+      const id_info = {
+        country: 'NG',
+        id_type: 'BUSINESS_REGISTRATION',
+        id_number: 'A000000',
+        business_type: '',
+      };
+
+      const scope = nock('https://testapi.smileidentity.com')
+        .post('/v1/business_verification', () => true)
+        .reply(400, businessVerificationResp.unsupported_business_type);
+
+      const promise = instance.submit_job(partner_params, id_info);
+      await expect(promise).rejects.toThrow(new Error('Request failed with status code 400'));
       expect(scope.isDone()).toBe(true);
     });
   });
